DPR-611 extract code handling athena query into a separate client class
