+++ backend/src/main/kotlin/uk/gov/justice/digital/backend/client/preview/PreviewClient.kt
@@ -1,0 +1,7 @@
+package uk.gov.justice.digital.backend.client.preview
+
+interface PreviewClient {
+
+    fun runQuery(query: String): List<Map<String, String>>
+
+}
